REorganizes Rust code
