+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "alignmentrs"
-version = "0.8.2"
+version = "0.8.3"
 authors = ["Kent Kawashima <kentkawashima@gmail.com>"]
 edition = "2018"
