+++ setup.py
@@ -22,13 +22,13 @@ setup(
     rust_extensions=[
         RustExtension('libalignmentrs.alignment',
                       'Cargo.toml', binding=Binding.PyO3),
-        RustExtension('libalignmentrs.fasta',
+        RustExtension('libalignmentrs.record',
                       'Cargo.toml', binding=Binding.PyO3),
     ],
     packages=find_packages(exclude=['contrib', 'docs', 'tests*']),
     package_data={
         'alignmentrs': ['lib/libalignmentrs/alignment.cpython-37m-darwin.so',
-                        'lib/libalignmentrs/fasta.cpython-37m-darwin.so',
+                        'lib/libalignmentrs/record.cpython-37m-darwin.so',
                         ]},
     install_requires=['blockrs', 'numpy'],
     zip_safe=False,  # Rust extensions are not zip safe, like C-extensions.
