+++ src/alignment.rs
@@ -1,7 +1,9 @@
 use pyo3::prelude::*;
 use pyo3::{PyObjectProtocol, exceptions};
 
-use crate::record::Record;
+use std::fs::File;
+use std::io::{BufReader, BufRead};
+use regex::Regex;
 
 #[pyclass(subclass)]
 #[derive(Clone)]
@@ -769,56 +771,88 @@ impl BaseAlignment {
     }
 }
 
-#[pymethods]
-impl Record {
-    #[new]
-    /// Creates a new Record object from sequence_id, sequence_description
-    /// and sequence_str.
-    fn __new__(obj: &PyRawObject, id: &str, description: &str, sequence_str: &str) -> PyResult<()> {
-        let sequence_str = String::from(sequence_str);
-        obj.init(|_| {
-            Record {
-                id: id.to_string(),
-                description: description.to_string(),
-                sequence: sequence_str.to_string(),
-            }
-        })
-    }
-    fn sequence_to_uint32(&self) -> Vec<i32> {
-        let uints: Vec<i32> = self.sequence.chars().map(|x| {x as usize} as i32 ).collect();
-        uints
-    }
+lazy_static! {
+    static ref WS: Regex = Regex::new(r"\s+").unwrap();
 }
 
-// Customizes __repr__ and __str__ of PyObjectProtocol trait
-#[pyproto]
-impl PyObjectProtocol for Record {
-    fn __repr__(&self) -> PyResult<String> {
-        let desc:String = match self.description.chars().count() {
-            x if x > 20 => {
-                let mut desc:String = self.description.char_indices()
-                                        .filter(|(i, _)| *i < 17)
-                                        .map(|(_, c)| c)
-                                        .collect();
-                desc.push_str("...");
-                desc
-            },
-            _ => self.description.clone()
-        };
-        Ok(format!("Record(id=\"{id}\", len={seq_len}, description=\"{desc}\")", id=self.id, seq_len=self.sequence.len(), desc=desc))
-    }
+#[pyfunction]
+/// fasta_file_to_basealignments(data_str)
+/// 
+/// Reads FASTA file and creates marker and sequence BaseAlignments.
+fn fasta_file_to_basealignments(path: &str, marker_kw: &str) -> 
+        PyResult<(BaseAlignment, BaseAlignment)> {
+    // Open the path in read-only mode, returns `io::Result<File>`
+    let f = match File::open(path) {
+        Err(x) => return Err(exceptions::IOError::py_err(format!("encountered an error while trying to open file {:?}: {:?}", path, x.kind()))),
+        Ok(x) => x
+    };
+    let f = BufReader::new(f);
 
-    fn __str__(&self) -> PyResult<String> {
-        if self.description.len() > 0 {
-            return Ok(format!(">{id} {desc}\n{seq_len}",
-                id=self.id,
-                desc=self.description,
-                seq_len=self.sequence.len()))
+    // Declare variables
+    let mut s_ids: Vec<String> = Vec::new();
+    let mut s_descriptions: Vec<String> = Vec::new();
+    let mut s_sequences: Vec<String> = Vec::new();
+
+    let mut m_ids: Vec<String> = Vec::new();
+    let mut m_descriptions: Vec<String> = Vec::new();
+    let mut m_sequences: Vec<String> = Vec::new();
+
+    let mut id = String::new();
+    let mut description = String::new();
+    let mut sequence = String::new();
+
+    // Match regexp
+    for line in f.lines() {
+        let line = match line {
+            Err(x) => return Err(exceptions::IOError::py_err(format!("encountered an error while reading file {:?}: {:?}", path, x.kind()))),
+            Ok(x) => x.trim().to_string()
+        };
+        if line.starts_with(">") {
+            if sequence.len() > 0 {
+                if marker_kw != "" && id.contains(marker_kw) {
+                    m_ids.push(id.clone());
+                    m_descriptions.push(description.clone());
+                    m_sequences.push(sequence.clone());
+                } else {
+                    s_ids.push(id.clone());
+                    s_descriptions.push(description.clone());
+                    s_sequences.push(sequence.clone());
+                }
+                sequence.clear();
+            }
+            let matches: Vec<&str> = WS.splitn(line.trim_start_matches(">"), 2).collect();
+            id = matches[0].to_string();
+            description = match matches.len() {
+                l if l == 2 => matches[1].to_string(),
+                _ => String::new(),
+            };
+        } else {
+            sequence.push_str(&line);
         }
-        return Ok(format!(">{id}\n{seq_len}",
-                id=self.id,
-                seq_len=self.sequence.len()))        
     }
+    if sequence.len() > 0 {
+        if marker_kw != "" && id.contains(marker_kw) {
+            m_ids.push(id);
+            m_descriptions.push(description);
+            m_sequences.push(sequence.clone());
+        } else {
+            s_ids.push(id);
+            s_descriptions.push(description);
+            s_sequences.push(sequence.clone());
+        }
+        sequence.clear();
+    }
+    let sample_aln = BaseAlignment {
+        ids: s_ids,
+        descriptions: s_descriptions,
+        sequences: s_sequences,
+    };
+    let marker_aln = BaseAlignment {
+        ids: m_ids,
+        descriptions: m_descriptions,
+        sequences: m_sequences,
+    };
+    Ok((sample_aln, marker_aln))
 }
 
 #[pyfunction]
@@ -843,13 +877,32 @@ fn concat_basealignments(aln_list: Vec<&BaseAlignment>) -> PyResult<BaseAlignmen
     Ok(BaseAlignment {ids, descriptions, sequences})
 }
 
+// Record
+
+#[pyclass(subclass)]
+#[derive(Clone)]
+/// Record(id, description, sequence)
+/// 
+/// Record represents a single sequence sample.
+pub struct Record {
+    #[prop(get,set)]
+    pub id: String,
+
+    #[prop(get,set)]
+    pub description: String,
+    
+    #[prop(get,set)]
+    pub sequence: String,
+
+}
+
 // Register python functions to PyO3
 #[pymodinit]
 fn alignment(_py: Python, m: &PyModule) -> PyResult<()> {
-    m.add_function(wrap_function!(concat_basealignments))?;
-
     m.add_class::<BaseAlignment>()?;
-    m.add_class::<Record>()?;
+    m.add_function(wrap_function!(fasta_file_to_basealignments))?;
+    m.add_function(wrap_function!(concat_basealignments))?;
+    m.add_function(wrap_function!(fasta_file_to_records))?;
 
     Ok(())
 }
