+++ src/lib.rs
@@ -5,7 +5,6 @@
 extern crate regex;
 
 pub mod alignment;
-pub mod fasta;
 
 // #[cfg(test)]
 // mod tests {
